# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Documentation",
        "version": "1.0"
    },
    "paths": {
        "/api/register": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Register a new user",
                "operationId": "8a56853624e025573120a09a4c75d468",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/AuthRegister"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "User registered successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "token": {
                                            "type": "string",
                                            "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJ1c2VyX2lkIjoxfQ.I2n-Dob8ht8OBWrqDqtTcZpZWYPcc6V-L0k9VAnUmY4"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "The email has already been taken."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/login": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Login an existing user",
                "operationId": "a3b306d14572d1f4bd6c064b3233e7b8",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/AuthLogin"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "User logged in successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "token": {
                                            "type": "string",
                                            "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJ1c2VyX2lkIjoxfQ.I2n-Dob8ht8OBWrqDqtTcZpZWYPcc6V-L0k9VAnUmY4"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Unauthorized"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/campaigns": {
            "get": {
                "tags": [
                    "Campaign"
                ],
                "summary": "Listar todas as campanhas",
                "description": "Retorna uma lista de campanhas",
                "operationId": "f63bfb37edf1acfd021c50961b734ec0",
                "responses": {
                    "200": {
                        "description": "Lista de campanhas retornada com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Campaigns"
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Erro ao buscar campanhas"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Campaign"
                ],
                "summary": "Criar uma nova campanha",
                "description": "Cria uma nova campanha com os dados fornecidos",
                "operationId": "272d9b28b56c8e18d5ef4c7440a05def",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "budget",
                                    "start_date",
                                    "end_date"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Campanha de Influência"
                                    },
                                    "budget": {
                                        "type": "number",
                                        "example": 10000.5
                                    },
                                    "start_date": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2024-01-01"
                                    },
                                    "end_date": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2024-12-31"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Campanha criada com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Campaigns"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Erro de validação dos dados"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/campaigns/{campaign_id}/influencers": {
            "post": {
                "tags": [
                    "Campaign Influencer"
                ],
                "summary": "Attach influencer to a campaign",
                "operationId": "3e69586ab4a943cb866e5d21e815c8b3",
                "parameters": [
                    {
                        "name": "campaign_id",
                        "in": "path",
                        "description": "ID of the campaign",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "influencer_id",
                        "in": "query",
                        "description": "ID of the influencer",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Influencer successfully added to the campaign",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Influencer added to campaign"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Campaign or Influencer not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/campaigns/{id}/influencers": {
            "get": {
                "tags": [
                    "Campaign Influencer"
                ],
                "summary": "Get all influencers of a campaign",
                "operationId": "828762b20686929d85711cf2077e4d38",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the campaign",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of influencers for the campaign",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Influencer"
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Campaign not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "AuthLogin": {
                "required": [
                    "email",
                    "password"
                ],
                "properties": {
                    "email": {
                        "type": "string",
                        "example": "user@example.com"
                    },
                    "password": {
                        "type": "string",
                        "example": "password123"
                    }
                },
                "type": "object"
            },
            "AuthRegister": {
                "required": [
                    "email",
                    "password",
                    "name"
                ],
                "properties": {
                    "email": {
                        "type": "string",
                        "example": "user@example.com"
                    },
                    "password": {
                        "type": "string",
                        "example": "password123"
                    },
                    "name": {
                        "type": "string",
                        "example": "John Doe"
                    }
                },
                "type": "object"
            },
            "Campaigns": {
                "title": "Campaigns",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "Campanha de Influência"
                    },
                    "budget": {
                        "type": "number",
                        "format": "float",
                        "example": 10000.5
                    },
                    "start_date": {
                        "type": "string",
                        "format": "date",
                        "example": "2024-01-01"
                    },
                    "end_date": {
                        "type": "string",
                        "format": "date",
                        "example": "2024-12-31"
                    }
                },
                "type": "object"
            },
...
            "Influencer": {
                "required": [
                    "id",
                    "name",
                    "email"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "John Doe"
                    },
                    "email": {
                        "type": "string",
                        "format": "email",
                        "example": "john.doe@example.com"
                    },
                    "followers": {
                        "type": "integer",
                        "example": 10000
                    },
                    "platform": {
                        "type": "string",
                        "example": "Instagram"
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Insira seu token Bearer para autenticação",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}